+++ scripts/test/SysConfig/ConfigurationCategoriesGet.t
@@ -83,8 +83,8 @@ my @Tests = (
                 DisplayName => 'All Settings',
                 Files       => [],
             },
-            OTRSFree => {
-                DisplayName => 'OTRS Free',
+            OTRS => {
+                DisplayName => 'OTRS',
                 Files       => [
                     'Calendar.xml', 'CloudServices.xml', 'Daemon.xml', 'Framework.xml', 'GenericInterface.xml',
                     'ProcessManagement.xml', 'Ticket.xml'
@@ -104,8 +104,8 @@ my @Tests = (
                 DisplayName => 'All Settings',
                 Files       => [],
             },
-            OTRSFree => {
-                DisplayName => 'OTRS Free',
+            OTRS => {
+                DisplayName => 'OTRS',
                 Files       => [
                     'Calendar.xml', 'CloudServices.xml', 'Daemon.xml', 'Framework.xml', 'GenericInterface.xml',
                     'ProcessManagement.xml', 'Ticket.xml'
@@ -129,8 +129,8 @@ my @Tests = (
                 DisplayName => 'All Settings',
                 Files       => [],
             },
-            OTRSFree => {
-                DisplayName => 'OTRS Free',
+            OTRS => {
+                DisplayName => 'OTRS',
                 Files       => [
                     'Calendar.xml', 'CloudServices.xml', 'Daemon.xml', 'Framework.xml', 'GenericInterface.xml',
                     'ProcessManagement.xml', 'Ticket.xml'