+++ scripts/test/SysConfig/ConfigurationTranslatedGet.t
@@ -36,7 +36,7 @@ my @Tests = (
     {
         Name          => 'Ticket::Frontend::AgentTicketQueue###SortBy::Default',
         ExpectedValue => {
-            'Category'    => 'OTRSFree',
+            'Category'    => 'OTRS',
             'IsInvisible' => '0',
             'Metadata'    => "ticket::frontend::agentticketqueue###sortby::default--- alter
 definiert die standardmäßig eingestellten sortierkriterien für alle in der queue-ansicht angezeigten queues."
@@ -45,7 +45,7 @@ definiert die standardmäßig eingestellten sortierkriterien für alle in der qu
     {
         Name          => 'Ticket::Frontend::AgentTicketQueue###Order::Default',
         ExpectedValue => {
-            'Category'    => 'OTRSFree',
+            'Category'    => 'OTRS',
             'IsInvisible' => 0,
             'Metadata'    => 'ticket::frontend::agentticketqueue###order::default--- auf
 definiert die standardmäßig eingestellten sortierkriterien für alle in der queue-ansicht angezeigten queues, nachdem nach priorität sortiert wurde.'
@@ -54,7 +54,7 @@ definiert die standardmäßig eingestellten sortierkriterien für alle in der qu
     {
         Name          => 'Ticket::Frontend::AgentTicketService###SortBy::Default',
         ExpectedValue => {
-            'Category'    => 'OTRSFree',
+            'Category'    => 'OTRS',
             'IsInvisible' => 0,
             'Metadata'    => 'ticket::frontend::agentticketservice###sortby::default--- alter
 definiert die standardmäßig eingestellten sortierkriterien für alle in der service-ansicht angezeigten services.'
@@ -63,7 +63,7 @@ definiert die standardmäßig eingestellten sortierkriterien für alle in der se
     {
         Name          => 'Ticket::Frontend::AgentTicketSearch###SearchCSVData',
         ExpectedValue => {
-            'Category'    => 'OTRSFree',
+            'Category'    => 'OTRS',
             'IsInvisible' => 0,
             'Metadata'    => 'ticket::frontend::agentticketsearch###searchcsvdata---
 - ticketnumber
@@ -95,7 +95,7 @@ daten die verwendet werden um das suchergebnis im csv-format zu exportieren.'
     {
         Name          => 'Ticket::Frontend::AgentTicketPhone###SplitLinkType',
         ExpectedValue => {
-            'Category'    => 'OTRSFree',
+            'Category'    => 'OTRS',
             'IsInvisible' => 0,
             'Metadata'    => 'ticket::frontend::agentticketphone###splitlinktype---
 direction: target
@@ -106,7 +106,7 @@ bestimmt den standard-linktyp für geteilte tickets im agentenbereich.'
     {
         Name          => 'Ticket::Frontend::HistoryTypes###000-Framework',
         ExpectedValue => {
-            'Category'    => 'OTRSFree',
+            'Category'    => 'OTRS',
             'IsInvisible' => 0,
             'Metadata'    => 'ticket::frontend::historytypes###000-framework---
 addnote: added note (%s).